+++ src/main/java/net/lopymine/ip/config/particle/ParticleAnimationType.java
@@ -1,6 +1,7 @@
 package net.lopymine.ip.config.particle;
 
 import com.mojang.serialization.Codec;
+import java.util.Locale;
 import net.minecraft.util.*;
 
 public enum ParticleAnimationType implements StringIdentifiable {
@@ -15,6 +16,6 @@ public enum ParticleAnimationType implements StringIdentifiable {
 
 	@Override
 	public String asString() {
-		return this.name().toLowerCase();
+		return this.name().toLowerCase(Locale.ROOT);
 	}
 }