+++ src/main/java/net/lopymine/ip/utils/EffectUtils.java
@@ -6,7 +6,7 @@ import net.minecraft.entity.effect.*;
 public class EffectUtils {
 
 	//? if <=1.20.1 {
-	public static Optional<Integer> mixColors(Iterable<StatusEffectInstance> effects) {
+	/*public static Optional<Integer> mixColors(Iterable<StatusEffectInstance> effects) {
 		int i = 0;
 		int j = 0;
 		int k = 0;
@@ -29,6 +29,6 @@ public class EffectUtils {
 			return Optional.of(ArgbUtils2.getArgb(255, i / l, j / l, k / l));
 		}
 	}
-	//?}
+	*///?}
 
 }