+++ src/main/java/net/lopymine/ip/utils/NbtUtils.java
@@ -7,15 +7,15 @@ import net.minecraft.entity.effect.StatusEffectInstance;
 import net.minecraft.item.*;
 
 //? if >=1.21 {
-/*import net.minecraft.component.DataComponentTypes;
+import net.minecraft.component.DataComponentTypes;
 import net.minecraft.component.type.*;
-*///?} else {
-import net.minecraft.entity.effect.StatusEffectInstance;
+//?} else {
+/*import net.minecraft.entity.effect.StatusEffectInstance;
 import net.minecraft.potion.Potion;
 import net.minecraft.registry.Registries;
 import net.minecraft.util.Identifier;
 import net.minecraft.nbt.*;
-//?}
+*///?}
 
 @ExtensionMethod(OptionalExtension.class)
 public class NbtUtils {
@@ -34,14 +34,14 @@ public class NbtUtils {
 
 		if (stack.isOf(Items.CROSSBOW)) {
 			//? if >=1.21 {
-			/*return Optional.ofNullable(stack.getComponents().get(DataComponentTypes.CHARGED_PROJECTILES))
+			return Optional.ofNullable(stack.getComponents().get(DataComponentTypes.CHARGED_PROJECTILES))
 					.map(ChargedProjectilesComponent::getProjectiles)
 					.filter((list) -> !list.isEmpty())
 					.map((list) -> getColorFromPotionContentsStack(list.get(0)))
 					.filter(Optional::isPresent)
 					.map(Optional::get);
-			*///?} else {
-			return Optional.ofNullable(stack.getNbt())
+			//?} else {
+			/*return Optional.ofNullable(stack.getNbt())
 					.to("ChargedProjectiles", NbtList.class)
 					.toEmpty(false)
 					.toFirst(NbtCompound.class)
@@ -49,10 +49,10 @@ public class NbtUtils {
 					.map(NbtUtils::getColorFromPotionNbt)
 					.filter(Optional::isPresent)
 					.map(Optional::get);
-			//?}
+			*///?}
 		}
 
-		if (/*? if >=1.21 {*/ /*stack.isIn(net.minecraft.registry.tag.ItemTags.DYEABLE) *//*?} else {*/ stack.getItem() instanceof DyeableItem /*?}*/ ) {
+		if (/*? if >=1.21 {*/ stack.isIn(net.minecraft.registry.tag.ItemTags.DYEABLE) /*?} else {*/ /*stack.getItem() instanceof DyeableItem *//*?}*/ ) {
 			return getColorFromDyedStack(stack);
 		}
 
@@ -69,15 +69,15 @@ public class NbtUtils {
 
 	public static Optional<Integer[]> getColorFromFirework(ItemStack stack) {
 		//? if >=1.21 {
-		/*return Optional.ofNullable(stack.getComponents().get(DataComponentTypes.FIREWORKS))
+		return Optional.ofNullable(stack.getComponents().get(DataComponentTypes.FIREWORKS))
 				.map(FireworksComponent::explosions)
 				.map((c) -> c.stream()
 						.map(FireworkExplosionComponent::colors)
 						.flatMap((o) -> o.intStream().boxed())
 						.toArray(Integer[]::new)
 				);
-		*///?} else {
-		return Optional.ofNullable(stack.getNbt())
+		//?} else {
+		/*return Optional.ofNullable(stack.getNbt())
 				.to("Fireworks", NbtCompound.class)
 				.to("Explosions", NbtList.class)
 				.toEmpty(false)
@@ -87,24 +87,24 @@ public class NbtUtils {
 						.toArray(Integer[]::new)
 				);
 
-		//?}
+		*///?}
 	}
 
 	public static Optional<Integer[]> getColorFromFireworkExplosionStack(ItemStack stack) {
 		//? if >=1.21 {
-		/*return Optional.ofNullable(stack.getComponents().get(DataComponentTypes.FIREWORK_EXPLOSION))
+		return Optional.ofNullable(stack.getComponents().get(DataComponentTypes.FIREWORK_EXPLOSION))
 				.map(NbtUtils::getColorFromFireworkExplosionStack);
-		*///?} else {
-		return Optional.ofNullable(stack.getNbt())
+		//?} else {
+		/*return Optional.ofNullable(stack.getNbt())
 				.to("Explosion")
 				.map(NbtUtils::getColorFromFireworkExplosionStack)
 				.filter(Optional::isPresent)
 				.map(Optional::get);
-		//?}
+		*///?}
 	}
 
 	//? if >=1.21 {
-	/*private static Integer[] getColorFromFireworkExplosionStack(FireworkExplosionComponent component) {
+	private static Integer[] getColorFromFireworkExplosionStack(FireworkExplosionComponent component) {
 		Integer[] colors = new Integer[component.colors().size()];
 
 		for (int i = 0; i < component.colors().size(); i++) {
@@ -113,8 +113,8 @@ public class NbtUtils {
 
 		return colors;
 	}
-	*///?} else {
-	public static Optional<Integer[]> getColorFromFireworkExplosionStack(NbtElement element) {
+	//?} else {
+	/*public static Optional<Integer[]> getColorFromFireworkExplosionStack(NbtElement element) {
 		return Optional.ofNullable(element)
 				.to(NbtCompound.class)
 				.to("Colors", NbtIntArray.class)
@@ -124,11 +124,11 @@ public class NbtUtils {
 						.map(NbtUtils::notZeroAlpha)
 						.toArray(Integer[]::new));
 	}
-	//?}
+	*///?}
 
 	public static Optional<Integer[]> getColorFromPotionContentsStack(ItemStack stack) {
 		//? if >=1.21 {
-		/*PotionContentsComponent component = stack.getComponents().get(DataComponentTypes.POTION_CONTENTS);
+		PotionContentsComponent component = stack.getComponents().get(DataComponentTypes.POTION_CONTENTS);
 		if (component != null) {
 			Optional<Integer> optional = component.customColor();
 			if (optional.isPresent()) {
@@ -147,32 +147,32 @@ public class NbtUtils {
 			return Optional.of(new Integer[]{-13083194});
 		}
 		return Optional.of(colors.toArray(Integer[]::new));
-		*///?} else {
-		return Optional.ofNullable(stack.getNbt())
+		//?} else {
+		/*return Optional.ofNullable(stack.getNbt())
 				.map(NbtUtils::getColorFromPotionNbt)
 				.filter(Optional::isPresent)
 				.map(Optional::get);
-		//?}
+		*///?}
 	}
 
 	public static Optional<Integer[]> getColorFromDyedStack(ItemStack stack) {
 		//? if >=1.21 {
-		/*return Optional.ofNullable(stack.getComponents().get(DataComponentTypes.DYED_COLOR))
+		return Optional.ofNullable(stack.getComponents().get(DataComponentTypes.DYED_COLOR))
 				.map(DyedColorComponent::rgb)
 				.map(NbtUtils::notZeroAlpha)
 				.map((i) -> new Integer[]{i});
-		*///?} else {
-		return Optional.ofNullable(stack.getNbt())
+		//?} else {
+		/*return Optional.ofNullable(stack.getNbt())
 				.to("display", NbtCompound.class)
 				.to("color", NbtInt.class)
 				.map(NbtInt::intValue)
 				.map(NbtUtils::notZeroAlpha)
 				.map((i) -> new Integer[]{i});
-		//?}
+		*///?}
 	}
 
 	//? if <=1.20.1 {
-	public static Optional<Integer[]> getColorFromPotionNbt(NbtCompound compound) {
+	/*public static Optional<Integer[]> getColorFromPotionNbt(NbtCompound compound) {
 		Optional<List<StatusEffectInstance>> optional = Optional.ofNullable(compound.get("Potion"))
 				.to(NbtString.class)
 				.map(NbtElement::asString)
@@ -186,7 +186,7 @@ public class NbtUtils {
 				.map(Optional::get)
 				.map((i) -> new Integer[]{i});
 	}
-	//?}
+	*///?}
 
 	private static int notZeroAlpha(int color) {
 		int alpha = ArgbUtils2.getAlpha(color);