+++ src/main/java/net/lopymine/ip/utils/RecursiveCodec.java
@@ -2,7 +2,7 @@ package net.lopymine.ip.utils;
 
 //? if <1.21 {
 
-import com.google.common.base.Suppliers;
+/*import com.google.common.base.Suppliers;
 
 import com.mojang.datafixers.util.Pair;
 import com.mojang.serialization.*;
@@ -35,4 +35,4 @@ public class RecursiveCodec<T> implements Codec<T> {
 	}
 }
 
-//?}
+*///?}